[refactor]: MonthlyRecordViewModel 삭제, DoItViewModel로 통합
